Alteration

+++ rest/api/database/models/bank_account.go
@@ -1,10 +1,10 @@
 package models
 
 type BankAccount struct {
-	ID         string `json:"id"`
-	Value      int    `json:"value"`
-	CurrencyId int    `json:"currencyId"`
-	UserId     int    `json:"userId"`
+	ID         string  `json:"id"`
+	Value      float64 `json:"value"`
+	CurrencyId int     `json:"currencyId"`
+	UserId     int     `json:"userId"`
 }
 
 type BankAccountList struct {
